requirements.txt
PyQt5
ffmpeg-python
send2trash 